brain_tumor/requirements.txt
opencv-python
tensorflow
pillow
imutils